requirements.txt
msal>=1.26.0
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
schedule>=1.2.1
colorlog>=6.8.0
fastapi>=0.100.0
uvicorn>=0.22.0
jinja2>=3.1.2
